+++ R/configurations.R
@@ -7,7 +7,7 @@
 #' Set the log file that loggit will write to. No logs will be written until
 #' this is set, as per CRAN policy. The suggested use of this function would be
 #' to call it early, to log to the current working directory, as follows:
-#' `setLogFile(paste0(getwd(), "/loggit.json"))`. If you are using `loggit` in
+#' `set_logfile(paste0(getwd(), "/loggit.json"))`. If you are using `loggit` in
 #' your package, you can wrap this function in `.onLoad()` so that the logfile
 #' is set when your package loads.
 #'
@@ -15,37 +15,33 @@
 #'   introduced, the file name must end in ".json".
 #' @param confirm Print confirmation of log file setting? Defaults to `TRUE`.
 #'
-#' @examples setLogFile(file.path(tempdir(), "loggit.json"))
+#' @examples set_logfile(file.path(tempdir(), "loggit.json"))
 #'
 #' @export
-setLogFile <- function(logfile = NULL, confirm = TRUE) {
-  
+set_logfile <- function(logfile = NULL, confirm = TRUE) {
   if (is.null(logfile)) {
     .config$logfile <- file.path(tempdir(), "loggit.json")
   } else {
-    
-    if (substr(logfile, nchar(logfile) - 4, nchar(logfile)) != ".json")
+    if (substr(logfile, nchar(logfile) - 4, nchar(logfile)) != ".json") {
       base::stop("Log file path must be explicitly JSON, i.e. end in '.json'")
+    }
     .config$logfile <- logfile
     .config$templogfile <- FALSE
     if (confirm) print(paste0("Log file set to ", logfile))
-    
   }
-  
-  invisible()
 }
+
 
 #' Get Log File
 #'
-#' Get the log file that loggit will write to.
+#' Return the log file that loggit will write to.
 #' 
-#' @examples getLogFile()
+#' @examples get_logfile()
 #'
 #' @export
-getLogFile <- function() {
+get_logfile <- function() {
   .config$logfile
 }
-
 
 
 #' Set Timestamp Format
@@ -56,22 +52,29 @@ getLogFile <- function() {
 #' @param confirm Print confirmation of timestamp format setting? Defaults to
 #'   `TRUE`.
 #'   
-#' @examples setTimestampFormat("%Y-%m-%d %H:%M:%S")
+#' @examples set_timestamp_format("%Y-%m-%d %H:%M:%S")
 #'
 #' @export
-setTimestampFormat <- function(ts_format = "%Y-%m-%d %H:%M:%S", confirm = TRUE) {
+set_timestamp_format <- function(ts_format = "%Y-%m-%d %H:%M:%S", confirm = TRUE) {
   .config$ts_format <- ts_format
   if (confirm) print(paste0("Timestamp format set to ", ts_format))
-  invisible()
 }
+
 
 #' Get Timestamp Format
 #' 
 #' Get timestamp format for use in output logs.
 #' 
-#' @examples getTimestampFormat()
+#' @examples get_timestamp_format()
 #'
 #' @export
-getTimestampFormat <- function() {
+get_timestamp_format <- function() {
   .config$ts_format
 }
+
+
+# Functions to be deprecated in next version
+setLogFile <- function()
+getLogFile
+setTimestampFormat
+getTimestampFormat