+++ R/loggit.R
@@ -40,7 +40,7 @@ loggit <- function(log_lvl, log_msg, log_detail = "", ..., echo = TRUE, custom_l
   
   if (.config$templogfile && !.config$seenmessage) {
     base::warning(paste0("loggit has no persistent log file. Please set with ",
-                         "setLogFile(logfile), or see package?loggit for more help.\n ",
+                         "set_logfile(logfile), or see package?loggit for more help.\n ",
                          "Otherwise, you can recover your logs (from THIS R SESSION ONLY) ",
                          "via copying ", .config$logfile, " to a persistent folder."))
     if (log_detail == "") log_detail <- "User was warned about non-persistent log file."
@@ -106,28 +106,25 @@ loggit <- function(log_lvl, log_msg, log_detail = "", ..., echo = TRUE, custom_l
 #' Return Log File as an R Object
 #'
 #' This function returns a `data.frame` (by default) containing all the logs in
-#' the provided JSON log file. If no explicit log fie is provided, calling this
-#' function will return a data frame of the log file currently pointed to by the
-#' loggit functions. Users can request that the logs be returned as a `list`,
-#' though this is not recommended.
+#' the provided ndJSON log file. If no explicit log file is provided, calling
+#' this function will return a data frame of the log file currently pointed to
+#' by the loggit functions.
 #'
-#' @param logfile JSON-format log file to return.
-#' @param as_df Should logfile be returned in a `data.frame` vs. a `list`?
-#'   Defaults to `TRUE`, and will return a `data.frame`.
+#' @param logfile ndJSON-format log file to return.
 #'
 #' @return A `data.frame`.
-#' 
+#'
 #' @examples
-#' setLogFile(file.path(tempdir(), "loggit.json"), confirm = FALSE)
+#' set_logfile(file.path(tempdir(), "loggit.json"), confirm = FALSE)
 #' message("Test log message")
-#' get_logs(getLogFile())
+#' read_logs(getLogFile())
 #'
 #' @export
-get_logs <- function(logfile, as_df = TRUE) {
+read_logs <- function(logfile, as_df = TRUE) {
   if (missing(logfile)) logfile <- .config$logfile
   if (!file.exists(logfile)) {
     base::stop("Log file does not exist")
   } else {
-    jsonlite::read_json(logfile, simplifyVector = as_df)
+    read_ndjson(logfile)
   }
 }